The New 90's Battle-City
This is the first time i've made and release a game. I made the game which adapted from the original game.The 1V1 modes have been added! But it does not support LAN connectivity .In addition to this. Of course you can try this game and can also give me advice.